BUG FIXES - MAJOR - Fixed bug where not all cities would show on main recycler view. Fixed bug where api would be called every time fragment came into view

+++ app/src/main/java/com/cseltz/android/weather/model/remote/remotedataclasses/weather/helpers/Hourly.kt
@@ -8,7 +8,7 @@ import kotlinx.parcelize.Parcelize
 data class Hourly(
     val clouds: Int,
     val dew_point: Double,
-    @SerializedName("dt") val time: Int,
+    @SerializedName("dt") val time: Long,
     val feels_like: Double,
     val humidity: Int,
     val pop: Double, // Probability of precipitation
